Another small BitSet fix.

+++ util/src/main/scala/com/precog/util/BitSetUtil.scala
@@ -24,7 +24,7 @@ import scala.annotation.tailrec
 object BitSetUtil {
   class BitSetOperations(bs: BitSet) {
     def toUnboxedArray(): Array[Long] = bitSetToArray(bs)
-    def toList(): List[Long] = bitSetToList(bs)
+    def toList(): List[Int] = bitSetToList(bs)
 
     def min(): Int = {
       val n = bs.nextSetBit(0)
@@ -110,9 +110,9 @@ object BitSetUtil {
     arr
   }
 
-  def bitSetToList(bs: BitSet): List[Long] = {
+  def bitSetToList(bs: BitSet): List[Int] = {
     @tailrec
-    def loopBits(long: Long, bit: Int, base: Int, sofar: List[Long]): List[Long] = {
+    def loopBits(long: Long, bit: Int, base: Int, sofar: List[Int]): List[Int] = {
       if (bit < 0)
         sofar
       else if (((long >> bit) & 1) == 1)
@@ -122,7 +122,7 @@ object BitSetUtil {
     }
 
     @tailrec
-    def loopLongs(i: Int, longs: Array[Long], base: Int, sofar: List[Long]): List[Long] = {
+    def loopLongs(i: Int, longs: Array[Long], base: Int, sofar: List[Int]): List[Int] = {
       if (i < 0)
         sofar
       else
